+++ data/src/main/java/com/mument_android/data/controller/LikeMumentController.kt
@@ -1,7 +1,7 @@
 package com.mument_android.data.controller
 
 import com.mument_android.data.dto.LikeCountDto
-import com.mument_android.core.base.BaseResponse
+import com.mument_android.data.util.BaseResponse
 
 interface LikeMumentController {
     suspend fun likeMument(mumentId: String, userId: String): BaseResponse<LikeCountDto>
