+++ data/src/main/java/com/mument_android/data/controller/RecordController.kt
@@ -1,7 +1,7 @@
 package com.mument_android.data.controller
 import com.mument_android.data.dto.record.MumentRecordDto
 import com.mument_android.data.dto.record.ResponseRecordMumentDto
-import com.mument_android.core.base.BaseResponse
+import com.mument_android.data.util.BaseResponse
 
 interface RecordController {
     suspend fun recordMument(
