+++ data/src/main/java/com/mument_android/data/controller/RecordControllerImpl.kt
@@ -3,7 +3,7 @@ package com.mument_android.data.controller
 import com.mument_android.data.dto.record.MumentRecordDto
 import com.mument_android.data.dto.record.ResponseRecordMumentDto
 import com.mument_android.data.network.record.RecordApiService
-import com.mument_android.core.base.BaseResponse
+import com.mument_android.data.util.BaseResponse
 import javax.inject.Inject
 
 class RecordControllerImpl @Inject constructor(
