+++ data/src/main/java/com/mument_android/data/controller/RecordModifyController.kt
@@ -3,7 +3,7 @@ package com.mument_android.data.controller
 
 import com.mument_android.data.dto.record.MumentRecordDto
 import com.mument_android.data.dto.record.ResponseRecordMumentDto
-import com.mument_android.core.base.BaseResponse
+import com.mument_android.data.util.BaseResponse
 
 interface RecordModifyController {
     suspend fun recordModifyMument(
