+++ data/src/main/java/com/mument_android/data/datasource/detail/MumentDetailDataSource.kt
@@ -1,7 +1,7 @@
 package com.mument_android.data.datasource.detail
 
 import com.mument_android.data.dto.detail.MumentDetailDto
-import com.mument_android.core.base.BaseResponse
+import com.mument_android.data.util.BaseResponse
 import kotlinx.coroutines.flow.Flow
 
 interface MumentDetailDataSource {
