+++ data/src/main/java/com/mument_android/data/datasource/detail/MumentDetailDataSourceImpl.kt
@@ -2,7 +2,7 @@ package com.mument_android.data.datasource.detail
 
 import com.mument_android.data.dto.detail.MumentDetailDto
 import com.mument_android.data.network.detail.DetailApiService
-import com.mument_android.core.base.BaseResponse
+import com.mument_android.data.util.BaseResponse
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flow
