+++ data/src/main/java/com/mument_android/data/datasource/detail/MumentListDataSource.kt
@@ -1,7 +1,7 @@
 package com.mument_android.data.datasource.detail
 
 import com.mument_android.data.dto.MumentListDto
-import com.mument_android.core.base.BaseResponse
+import com.mument_android.data.util.BaseResponse
 
 interface MumentListDataSource {
     suspend fun fetchMumentList(
