+++ data/src/main/java/com/mument_android/data/datasource/detail/MumentListDataSourceImpl.kt
@@ -2,7 +2,7 @@ package com.mument_android.data.datasource.detail
 
 import com.mument_android.data.dto.MumentListDto
 import com.mument_android.data.network.detail.DetailApiService
-import com.mument_android.core.base.BaseResponse
+import com.mument_android.data.util.BaseResponse
 import javax.inject.Inject
 
 class MumentListDataSourceImpl @Inject constructor(
