+++ data/src/main/java/com/mument_android/data/datasource/detail/MusicDetailDataSource.kt
@@ -1,7 +1,7 @@
 package com.mument_android.data.datasource.detail
 
 import com.mument_android.data.dto.detail.MusicDetailDto
-import com.mument_android.core.base.BaseResponse
+import com.mument_android.data.util.BaseResponse
 
 interface MusicDetailDataSource {
     suspend fun fetchMusicDetailInfo(musicId: String, userId: String): BaseResponse<MusicDetailDto>
