+++ data/src/main/java/com/mument_android/data/datasource/home/HomeDataSource.kt
@@ -4,7 +4,7 @@ import com.mument_android.data.dto.home.BannerMumentDto
 import com.mument_android.data.dto.home.KnownMumentDto
 import com.mument_android.data.dto.home.RandomMumentDto
 import com.mument_android.data.dto.home.TodayMumentDto
-import com.mument_android.core.base.BaseResponse
+import com.mument_android.data.util.BaseResponse
 
 interface HomeDataSource {
     suspend fun getBannerMument(): BaseResponse<BannerMumentDto>
