+++ data/src/main/java/com/mument_android/data/datasource/home/RemoteMumentHistoryDataSourceImpl.kt
@@ -2,7 +2,7 @@ package com.mument_android.data.datasource.home
 
 import com.mument_android.data.dto.history.MumentHistoryDto
 import com.mument_android.data.network.home.HomeService
-import com.mument_android.core.base.BaseResponse
+import com.mument_android.data.util.BaseResponse
 import javax.inject.Inject
 
 class RemoteMumentHistoryDataSourceImpl @Inject constructor(val homeService:HomeService):
