+++ data/src/main/java/com/mument_android/data/datasource/home/RemoteSearchListDataSource.kt
@@ -1,7 +1,7 @@
 package com.mument_android.data.datasource.home
 
 import com.mument_android.domain.entity.home.RecentSearchData
-import com.mument_android.core.base.BaseResponse
+import com.mument_android.data.util.BaseResponse
 
 interface RemoteSearchListDataSource {
     suspend fun searchMusicList(keyword:String): BaseResponse<List<RecentSearchData>>
