+++ data/src/main/java/com/mument_android/data/datasource/locker/LockerDataSource.kt
@@ -1,7 +1,7 @@
 package com.mument_android.data.datasource.locker
 
 import com.mument_android.data.dto.locker.LockerMyMumentDto
-import com.mument_android.core.base.BaseResponse
+import com.mument_android.data.util.BaseResponse
 
 interface LockerDataSource {
     suspend fun fetchLockerMumumentList(userId: String, tag1 : Int?, tag2: Int?, tag3: Int?): BaseResponse<LockerMyMumentDto>
