+++ data/src/main/java/com/mument_android/data/datasource/locker/LockerDataSourceImpl.kt
@@ -2,7 +2,7 @@ package com.mument_android.data.datasource.locker
 
 import com.mument_android.data.dto.locker.LockerMyMumentDto
 import com.mument_android.data.network.locker.LockerApiService
-import com.mument_android.core.base.BaseResponse
+import com.mument_android.data.util.BaseResponse
 import javax.inject.Inject
 
 class LockerDataSourceImpl @Inject constructor(
