+++ data/src/main/java/com/mument_android/data/network/locker/LockerApiService.kt
@@ -1,6 +1,6 @@
 package com.mument_android.data.network.locker
 
-import com.mument_android.core.base.BaseResponse
+import com.mument_android.data.util.BaseResponse
 import com.mument_android.data.dto.locker.LockerMyMumentDto
 import retrofit2.http.GET
 import retrofit2.http.Path
