+++ data/src/main/java/com/mument_android/data/network/main/MainApiService.kt
@@ -1,6 +1,6 @@
 package com.mument_android.data.network.main
 
-import com.mument_android.core.base.BaseResponse
+import com.mument_android.data.util.BaseResponse
 import com.mument_android.data.dto.LikeCountDto
 import retrofit2.http.DELETE
 import retrofit2.http.POST
