+++ feature/detail/src/main/java/com/mument_android/detail/util/SortTypeEnum.kt
@@ -1,4 +1,4 @@
-package com.mument_android.core.util
+package com.mument_android.detail.util
 
 enum class SortTypeEnum(val sort: String, val tag: String) {
     SORT_LIKE_COUNT("좋아요순", "Y"),
